Using externall class for startpoint
