+++ package.json
@@ -11,15 +11,15 @@
     "lint": "tslint --project tsconfig.json --format stylish"
   },
   "dependencies": {
-    "koa": "^2.11.0",
-    "koa-bodyparser": "^4.2.1",
-    "koa-router": "^8.0.6",
     "body-parser": "^1.19.0",
     "class-transformer": "^0.2.3",
     "class-validator": "^0.11.0",
     "express": "^4.17.1",
+    "koa": "^2.11.0",
+    "koa-bodyparser": "^4.2.1",
     "koa-logger": "^3.2.1",
     "koa-multer": "^1.0.2",
+    "koa-router": "^8.0.6",
     "mongodb": "^3.5.2",
     "multer": "^1.4.2",
     "routing-controllers": "^0.8.0",
@@ -31,19 +31,18 @@
     "@types/http-status-codes": "^1.2.0",
     "@types/koa": "^2.11.0",
     "@types/koa-bodyparser": "^4.3.0",
+    "@types/koa-logger": "^3.1.1",
     "@types/koa-router": "^7.4.0",
     "@types/mocha": "^7.0.1",
-    "@types/node": "^13.7.0",
-    "@types/koa-logger": "^3.1.1",
     "@types/mongodb": "^3.3.15",
     "@types/mongoose": "^5.7.0",
+    "@types/node": "^13.7.0",
     "@types/request": "^2.48.4",
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
     "http-status-codes": "^1.4.0",
     "mocha": "^7.0.1",
     "nodemon": "^2.0.2",
-    "pg": "^7.18.1",
     "reflect-metadata": "^0.1.13",
     "request": "^2.88.0",
     "ts-node": "^8.6.2",
